ビルドエラー解消対応１

+++ DB/DBZ/standard/business/src/test/java/jp/co/ndensan/reams/db/dbz/business/core/basic/ShisetsuNyutaishoIdentifierTest.java
@@ -10,6 +10,7 @@ import jp.co.ndensan.reams.db.dbz.entity.basic.helper.DbT1004ShisetsuNyutaishoEn
 import jp.co.ndensan.reams.db.dbz.testhelper.DbzTestBase;
 
 import jp.co.ndensan.reams.uz.uza.biz.ShikibetsuCode;
+import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import org.junit.BeforeClass;
@@ -26,7 +27,7 @@ public class ShisetsuNyutaishoIdentifierTest extends DbzTestBase {
 //TODO 主キー型と変数名を置換してください
 //TODO 主キーの数が足りない場合、追加してください。
     private static ShikibetsuCode 識別コード;
-    private static int 履歴番号;
+    private static Decimal 履歴番号;
 
     @BeforeClass
     public static void setUpClass() {
